+++ src/Concerns/ProvidesBrowser.php
@@ -29,6 +29,7 @@ trait ProvidesBrowser
      * Tear down the Dusk test case class.
      *
      * @afterClass
+     *
      * @return void
      */
     public static function tearDownDuskClass()