+++ src/Http/Controllers/UserController.php
@@ -32,7 +32,7 @@ class UserController
      * Login using the given user ID / email.
      *
      * @param  string  $userId
-     * @param  string  $guard
+     * @param  string|null  $guard
      * @return void
      */
     public function login($userId, $guard = null)
@@ -51,14 +51,16 @@ class UserController
     /**
      * Log the user out of the application.
      *
-     * @param  string  $guard
+     * @param  string|null  $guard
      * @return void
      */
     public function logout($guard = null)
     {
-        Auth::guard($guard ?: config('auth.defaults.guard'))->logout();
+        $guard = $guard ?: config('auth.defaults.guard');
 
-        Session::forget('password_hash_'.$guard ?: config('auth.defaults.guard'));
+        Auth::guard($guard)->logout();
+
+        Session::forget('password_hash_'.$guard);
     }
 
     /**