+++ tests/Concerns/InteractsWithElementsTest.php
@@ -77,9 +77,10 @@ class InteractsWithElementsTest extends TestCase
     /**
      * @covers ::value
      * @dataProvider dataProviderValueWithValue
-     * @param mixed  $selector
-     * @param mixed  $value
-     * @param string  $js
+     *
+     * @param  mixed  $selector
+     * @param  mixed  $value
+     * @param  string  $js
      */
     public function testValueWithValue($selector, $value, string $js)
     {